version 0.0.4

+++ app/src/main/java/com/kappdev/notes/feature_notes/data/data_source/NoteDao.kt
@@ -17,5 +17,5 @@ interface NoteDao {
     fun getNoteById(id: Long): Note
 
     @Delete
-    suspend fun removeNote(note: Note): Long
+    suspend fun removeNote(note: Note): Int
 }
